% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_for_ML.R
\name{cat_parallel}
\alias{cat_parallel}
\title{cat_parallel (from nestedcv)}
\usage{
cat_parallel(...)
}
\arguments{
\item{\dots}{to be passed to system()}
}
\description{
Prints using shell echo from inside mclapply when run in Rstudio
}
