% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_diffNetworks}
\alias{get_diffNetworks}
\title{Generate multi-omic differential networks}
\usage{
get_diffNetworks(
  assayData,
  metadata,
  category_variable = NULL,
  regression_method = "rlm",
  category_subset = NULL,
  network = NULL,
  entrezIDs = FALSE,
  convert_to_gene_symbols = FALSE,
  use_qvalues = FALSE,
  show_progressBar = TRUE,
  verbose = TRUE,
  cores = parallel::detectCores()/2
)
}
\arguments{
\item{assayData}{a matrix or data.frame (or list of matrices or data.frames
for multi-omic analysis) containing normalised assay data.
Sample IDs must be in columns and probe IDs (genes, proteins...) in rows.
For multi omic analysis, it is highly recommended to use a named list of data.
If unnamed, sequential names (assayData1, assayData2, etc.) will be
assigned to identify each matrix or data.frame.}

\item{metadata}{a named vector, matrix, or data.frame containing sample
annotations or categories. If matrix or data.frame, each row should
correspond to a sample, with columns representing different sample
characteristics (e.g., treatment group, condition, time point). The colname
of the sample characteristic to be used for differential analysis must be
specified in \code{category_variable}. Rownames must match the sample IDs used in
assayData.
If named vector, each element must correspond to a sample characteristic to
be used for differential andlysis, and names must match sample IDs used in
the colnames of \code{assayData}.
Continuous variables are not allowed.}

\item{category_variable}{when metadata is a matrix or data.frame this is the
column name of \code{metadata} that contains the sample annotations to be used for
differential analysis}

\item{regression_method}{whether to use robust linear modelling to calculate
link p values. Options are 'rlm' (default) or 'lm'.}

\item{category_subset}{optional character vector indicating a subset of
categories from the category variable. If not specified, all categories in
\code{category_variable} will be used.}

\item{network}{network of biological interactions provided by the user. The
network must be provided in the form of a table of class data.frame with only
two columns named "from" and "to".
If NULL (default) a network of 10,537 molecular interactions obtained from
KEGG, mirTARbase, miRecords and transmiR will be used.
This has been obtained via the \code{exportgraph} function of the MITHrIL tool
(Alaimo et al., 2016).}

\item{entrezIDs}{logical (default FALSE) when genes are in \code{assayData} are
entrez IDs (TRUE) or gene symbols (FALSE).}

\item{convert_to_gene_symbols}{logical to be used when assayData entrez IDs
must be converted to gene symbols. Default is FALSE.}

\item{use_qvalues}{whether to use Storey's q values for multiple test
adjustment. If FALSE (default), unadjusted p values will be used and shown
in the output.}

\item{cores}{number of cores to use for parallelisation.}
}
\value{
a \code{deggs} object containing differential networks incorporating
p values or adjusted p values for each link.
}
\description{
Generate a multi-layer differential network with interaction p values
}
