% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_diffNetworks_singleOmic}
\alias{get_diffNetworks_singleOmic}
\title{Generate differential networks for single onimc analysis}
\usage{
get_diffNetworks_singleOmic(
  assayData,
  assayDataName,
  metadata,
  regression_method = "rlm",
  network = NULL,
  percentile_vector = seq(0.55, 0.98, by = 0.05),
  use_qvalues = FALSE,
  show_progressBar = TRUE,
  verbose = TRUE,
  cores = parallel::detectCores()/2
)
}
\arguments{
\item{assayData}{a matrix or data.frame with normalised assay data.
Sample IDs must be in columns and probe IDs (genes, proteins...) in rows.}

\item{assayDataName}{name of the assayData, to identify which omic is.}

\item{metadata}{a named vector, each element must correspond to a sample
characteristic to be used for differential analysis, and names must match
sample IDs used in the colnames of \code{assayData}.
Continuous variables are not allowed.}

\item{regression_method}{whether to use robust linear modelling to calculate
link p values. Options are 'rlm' (default) or 'lm'.}

\item{network}{network of biological interactions provided by the user. The
network must be provided in the form of a table of class data.frame with only
two columns named "from" and "to".
If NULL (default) a network of 10,537 molecular interactions obtained from
KEGG, mirTARbase, miRecords and transmiR will be used.
This has been obtained via the \code{exportgraph} function of the MITHrIL tool
(Alaimo et al., 2016).}

\item{use_qvalues}{whether to use Storey's q values for multiple test
adjustment. If FALSE (default), unadjusted p values will be used and shown
in the output.}

\item{cores}{number of cores to use for parallelisation.}
}
\value{
a list of differential networks, one per category
}
\description{
Generate differential networks for single onimc analysis
}
