% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_sig_deggs}
\alias{get_sig_deggs}
\title{Get a table of all the significant interactions across categories}
\usage{
get_sig_deggs(deggs_object, assayDataName = 1, sig_threshold = 0.05)
}
\arguments{
\item{deggs_object}{an object of class \code{deggs} generated by
\code{get_diffNetworks}}

\item{assayDataName}{name or index of the assayData of interest.
Accepts either the name of the assayData (e.g. \code{assayDataName = "RNAseq"})
or its position in the list as an integer
(e.g. \code{assayDataName = 2} for the second element). Default 1.}

\item{sig_threshold}{threshold for significance. Default 0.05.}
}
\value{
a \code{data.frame} listing all the significant differential interactions
found across categories for that particular omic data.
This list can also be used to substitute or integrate feature selection in
machine learning models for the prediction of the categories (see vignette).
}
\description{
Get a table of all the significant interactions across categories
}
\examples{
data("synthetic_metadata")
data("synthetic_rnaseqData")
deggs_object <- get_diffNetworks(assayData = synthetic_rnaseqData,
                                 metadata = synthetic_metadata,
                                 category_variable = "response",
                                 verbose = FALSE,
                                 show_progressBar = FALSE,
                                 cores = 2)
get_sig_deggs(deggs_object, sig_threshold = 0.05) 
}
