% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{my_palette}
\alias{my_palette}
\title{Internal function for colors}
\usage{
my_palette(n)
}
\arguments{
\item{n}{number of colors needed}
}
\value{
a vector with colors
}
\description{
This function return a color palette with the number of colors specified by n
}
