% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{node_boxplot}
\alias{node_boxplot}
\title{Boxplots of single nodes (genes,proteins, etc.)}
\usage{
node_boxplot(gene, assayDataName = 1, deggs_object)
}
\arguments{
\item{gene}{gene name  (must be in \code{rownames(assayData)})}

\item{assayDataName}{name of the assayData of interest. If an unnamed list of
data was given to \code{get_diffNetworks}, the assayDataName here will be the
number indicating the position of the data in the assayDataList provided
before (i.e. if the user wants to plot a differential interaction observed in
the transcriptomic data, which was second in the list, then assayDataName
must be 2, if only one data table was provided assayDataName must be 1).
Default 1.}

\item{deggs_object}{an object of class \code{deggs} generated by
\code{get_diffNetworks}}
}
\value{
the boxplot
}
\description{
This function is for internal use of \code{View_diffnetworks}
}
