% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{synthetic_metadata}
\alias{synthetic_metadata}
\title{Synthetic clinical data}
\format{
A data frame with 100 rows and 4 columns (IDs are in rownames):
\describe{
\item{response}{The response outcome, to be used for differential analysis}
}
}
\usage{
synthetic_metadata
}
\description{
A dataset containing sample clinical data for 100 patients with 40\%
response rate
}
\keyword{datasets}
