% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{synthetic_proteomicData}
\alias{synthetic_proteomicData}
\title{Synthetic RNA-seq count data}
\format{
A data frame with xx rows (proteins) xx columns (patients IDs).
}
\description{
Synthetic RNA-seq data after log2 normalisation
}
\keyword{datasets}
