% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{synthetic_rnaseqData}
\alias{synthetic_rnaseqData}
\title{Synthetic RNA-seq count data}
\format{
A data frame with xx rows (genes) xx columns (patients IDs, matching
the metadata rownames).
}
\description{
Synthetic RNA-seq data after log2 normalisation
}
\keyword{datasets}
