% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{tidy_metadata}
\alias{tidy_metadata}
\title{Tidying up of metadata and id_variable (when provided).}
\usage{
tidy_metadata(
  category_subset,
  metadata,
  category_variable,
  id_variable,
  verbose
)
}
\arguments{
\item{category_subset}{optional character vector indicating a subset of
categories from the category variable. If not specified, all categories in
\code{category_variable} will be used.}

\item{metadata}{a named vector, matrix, or data.frame containing sample
annotations or categories. If matrix or data.frame, each row should
correspond to a sample, with columns representing different sample
characteristics (e.g., treatment group, condition, time point). The colname
of the sample characteristic to be used for differential analysis must be
specified in \code{category_variable}. Rownames must match the sample IDs used in
assayData.
If named vector, each element must correspond to a sample characteristic to
be used for differential analysis, and names must match sample IDs used in
the colnames of \code{assayData}.
Continuous variables are not allowed.}

\item{category_variable}{when metadata is a matrix or data.frame this is the
column name of \code{metadata} that contains the sample annotations to be used for
differential analysis}

\item{id_variable}{character. Name of the column in metadata containing the
grouping variable for random effects (e.g., "slide_id", "batch"). Only used
when mixedModel = TRUE. Default NULL.}

\item{verbose}{logical. Whether to print detailed output messages during
processing. Default is TRUE}
}
\value{
a tidy named factor vector of sample annotations,
or a list with metadata and id_vector when mixedModel = TRUE
}
\description{
Samples belonging to undesidered categories
(if specified) will be removed as well as categories with less than five
samples, and NAs.
Both metadata and id_variable (when provided) will be returned as named factors.
Used internally by get_diffNetworks_singleOmic.
}
\keyword{internal}
