% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{View_diffNetworks}
\alias{View_diffNetworks}
\title{Interactive visualisation of differential networks}
\usage{
View_diffNetworks(deggs_object, host = NULL, port = NULL)
}
\arguments{
\item{deggs_object}{an object of class \code{deggs} generated by
\code{get_diffNetworks}}

\item{host}{local IP address, to be used only to run the function in
non-blocking mode (see vignette)}

\item{port}{TCP port, to be used only to run the function in
non-blocking mode (see vignette)}
}
\value{
a shiny interface showing networks with selectable nodes and links
}
\description{
Explore differential networks and interactively select regression
and box plots
}
