% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{calc_pvalues_network}
\alias{calc_pvalues_network}
\title{Calculate the pvalues for specific category network samples}
\usage{
calc_pvalues_network(
  assayData,
  metadata,
  sig_var,
  categories_length,
  regression_method = "rlm",
  category_network
)
}
\arguments{
\item{assayData}{a matrix or data.frame (or list of matrices or data.frames
for multi-omic analysis) containing normalised assay data.
Sample IDs must be in columns and probe IDs (genes, proteins...) in rows.
For multi omic analysis, it is highly recommended to use a named list of data.
If unnamed, sequential names (assayData1, assayData2, etc.) will be
assigned to identify each matrix or data.frame.}

\item{metadata}{a named vector, matrix, or data.frame containing sample
annotations or categories. If matrix or data.frame, each row should
correspond to a sample, with columns representing different sample
characteristics (e.g., treatment group, condition, time point). The colname
of the sample characteristic to be used for differential analysis must be
specified in \code{category_variable}. Rownames must match the sample IDs used in
assayData.
If named vector, each element must correspond to a sample characteristic to
be used for differential andlysis, and names must match sample IDs used in
the colnames of \code{assayData}.
Continuous variables are not allowed.}

\item{sig_var}{Inherited from \code{get_diffNetworks}. It can be
\code{q.value} or \code{p.value} depending on how \code{use_qvalues} was set in the
\code{get_diffNetworks} function (default \code{FALSE}).}

\item{categories_length}{integer number indicating the number of categories}

\item{regression_method}{whether to use robust linear modelling to obtain
p value of the interactions. Options are 'rlm' (default) or 'lm'}

\item{category_network}{network table for a specific category}
}
\value{
a list of p values
}
\description{
Calculate the pvalues for specific category network samples
}
