% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_multiOmics_diffNetworks}
\alias{get_multiOmics_diffNetworks}
\title{Get a table of all significant interactions across categories}
\usage{
get_multiOmics_diffNetworks(deggs_object, sig_threshold = 0.05)
}
\arguments{
\item{deggs_object}{an object of class \code{deggs} generated by
\code{get_diffNetworks}}

\item{sig_threshold}{threshold for significance. Default 0.05.}
}
\value{
a list of multilayer networks (in form of edge table), one per
category in the data.
}
\description{
Get a table of all significant interactions across categories
}
