% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{synthetic_metadata}
\alias{synthetic_metadata}
\title{Synthetic clinical data}
\format{
A data frame with 100 rows and 4 columns (IDs are in rownames):
\describe{
\item{patient_id}{IDs matching the IDs used in the colnames of the assay
data matrix/matrices.}
\item{age}{A column to simulate age of patients. Not used.}
\item{gender}{A column to simulate gender of patients. Not used.}
\item{response}{The response outcome, to be used for differential analysis}
}
}
\description{
A dataset containing sample clinical data for 100 patients with 40\%
response rate
}
\keyword{datasets}
